Point lights seem to work.
The problem was that we had two situations.
Let V be the eye position, L be the light position
and let P be the position of the surface to light.

Then there are two situations:

V <--> L <--> P

That is, the light is inbetween the eye and the surface, and

L <--> V <--> P

That is, the eye, possibly looking at the surface, is behind
the light. The problem was that if we did

light_dir = light.position - position

then only the second situation would be lighted correctly,
while if we did

light_dir = position - light.position

then only the first situation would be lighted correctly.
The bug was in the calculation of the light contribution
(the lambert contribution). Here we did

max(dot(light_dir, normal), 0.0f)

but what actually should happen is

abs(dot(light_dir, normal))

This accounts for both situations correctly. With this commit,
point lights almost seem to work. The only problem is that if
we zoom in or out, the light seems to dim or enhance with our
own position. This should not happen; the eye position should
not affect the lighting on the surface of an object. I'll be
investigating further into this issue.