Changed the "intensity" property of lights to have a 4th component.
The fourth component (alpha, or w, or whatever) gives us a way to
multiply the final light intensity with a given factor. I added this
feature after seeing that point lights were kind of weak, even at full
intensity. So, the x, y and z components of the intensity value define
the color of the light. The w component defines how bright the light shines.
A good default value is 1 for directional lights. For point lights, you might
want to use a value of between 1 (kind of weak) to 10 (medium intensity) to
20 (very intense). Of course, this is all relative.