Added spatial tree datastructures.
The spatial tree datastructures are located in spatialtrees.hpp
There are currently two trees, one for two-dimensional data, the
quadtree, and one for three-dimensional data, the octree. The user
can query an area and the tree will return, in O(log n) time, the
objects located in that area (a rectangle or a cuboid).

The actor class has been adjusted so that it will work in tandem
with the new spatial datastructures. The idea is that actors will
be populating a spatial tree, so that area queries on which actors
need to be rendered are quickly computed when there are thousands of
actors.