Greatly improved the way the project is built.
The examples do not depend on what packages gintonic includes
now. So the CMakeLists file of the examples is more self-contained.
Furthermore, most of the package handling is now in the CMakeLists
file inside the lib folder (the folder where all of the sources of
gintonic reside). The top-level CMakeLists file has been stripped
of a lot of nonsense.

I am not sure if this will compile on Windows but we'll see.