Made point lights FINALLY work correctly.
The problemw was that the GBUFFER_POSITION color attachment
was GL_RGB. Instead, it MUST be GL_RGB32F in order to correctly
store the position information. If it's GL_RGB, then the values
are clamped to incorrect values. If it's GL_RGB32F, negative values
will be correctly stored. I'm not sure how this will work out in
Windows, because the last time I tried to compile with GL_RGB32F
it didn't work so well.