Fixed spot lights.
See examples/spot_lights.cpp for an example.

Optimization opportunity: During the stencil pass
of the spot light, a sphere is used for the stencil
test (just as with point lights). However, a spot light
naturally has more of a cone as a representation of its
light volume. There's a cone in basic_shapes.hpp but I
don't think it works correctly.