Made a *good* unit_sphere_PUN and a *decent* unit_cone_PN
The normal vectors on the unit_cone_PN could still use some
work. Also, the cones are rendered with triangle fans. Not
sure if this is a good idea. The problem is that the culling
of the triangles is reversed inside the draw call (from GL_CCW
to GL_CW) and then changed back to, what we assume, is the original
(namely GL_CCW). However we can't make that assumption. It works
fine for now but maybe needs more thought.