Added a new example called "normal_map_test"
This example should show how a normal texture
influences the normal vectors on a flat surface.

However it seems that my calculations for the tangent
and bitangent vectors are not yet quite correct. Some
faces of the cube in the example are "wrong", so to speak.
That is, the effect you should see is that there's a bump
that extrudes on each face, while right now on some (or most)
faces, it actually goes "into" the surface instead of "out of"
the surface.