Modified the point lights example to have more flexibility.
As it turns out, having an attenuation value of (0,0,1) is a
bad idea, since the cutoff value for the stencil sphere will
sometimes spazz out. This is because the discriminant of the
cutoff calculation becomes negative. Anyway, a "better" thing
to do is to start with an attenuation of (1,1,1), and increase
or decrease the quadratic component as one sees fit for the scene.