Constructing a gintonic::mesh from an FbxMesh works.
As a bonus, the algorithm to convert an FbxMesh to a
gintonic::mesh is a tiny bit smart now, because it checks
for duplicate vertices. If so the index array just references
that duplicate vertex. As a consequence, loading an FbxMesh
takes about O(n^2) time, where n is the number of vertices.
It used to be O(n) time, but we save a lot of memory now, and
no extra overhead during rendering.