Modified profiler class and driver info tool.
The profiler class lives in a namespace called
gintonic::detail now. Also using the __PRETTY_FUNCTION__
macro so that we don't have to supply the name of the
function anymore.

The driver info tool now only prints the renderer
name and version by default. If given the command line argument
"--extensions", or "-e", the tool will also print all the
OpenGL extensions of the renderer.