Started experimenting with a generic shader
and lua bindings. Currently trying out "Selene"
from GitHub, but it doesn't seem to support
more than one constructor. Maybe this is a limitation
of the Lua language itself, but I'm not sure at this
point. Perhaps I will have to manually write the lua
bindings myself. I think it would be a good exercise
to first experiment with classes inside Lua, creating
metatables and whatnot.