Tweaked the SpotLightShadowBuffer.
It seems to work fine now. The only "magic number" still present
is the choice of the near plane for a SpotLightShadowBuffer.
I do not know what kind of value it should take. In a theoretical
sense, it should be the distance from the closest triangle of
the closest mesh. But I do not know how I would compute that in
real time.