Found subtle bug involving alignment issues.
When making new Entities et al. with std::make_shared,
the function uses its own malloc function to allocate
memory and put the allocated object tightly beside the
control block. However, this means that that newly allocated
chunk is not necessarily 16-byte aligned. It was frustrating to
find this.

Another problem remains:

The unit test for materials is reporting memory leaks. I do not
know how that can happen.

The destructor for Entity seems to crash on Windows. I do not
know how that can happen.

These two issues could be related.