Fixed a bug in the constructor of the SQT class.
The bug was that the affine transformation matrix of an FbxNode
was not translated correctly.

However, now the FbxNode is const_casted to a mutable FbxNode,
the local transformation matrix is evaluated (this is why a
mutable one is needed), and then that transformation is used.

What should happen is that there should be a mutable constructor
for the SQT class, reflecting that the FbxNode is probably going
to be mutated. However this change will propagate through the
code base so it's going to be a bit of work.