Getting there with animation.
It's still extremely glitchy, but I am seeing movement of
vertices which are driven by animation clips.

New classes are Skeleton, Skeleton::Joint and AnimationClip.

The Mesh class has become "dumber", in the sense that it forgot
its bones. The Entity class now houses a vector of AnimationClips.

Each AnimationClip has a shared pointer to a Skeleton.

The Skeleton houses the Joints (formally known as Mesh::Bone).

All this comes together in the Renderer class. Specifically,
the static method Renderer::renderGeometry uses all of this
information to do animation.