Replaced freetype as a git submodule.
Right now, the only "find_package" commands left in cmake are
Boost, OpenGL and FBX. For FBX, it's impossible to add it as a
git submodule. For boost, I think there was some effort to turn
all libraries into cmake-compliant libs, but I don't think there is
much animo for that. OpenGL libraries are standard, so it's find
to have that as a "find_package" command.

So users will have to do two things now: download boost and download
FBX.